Create a function for division

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -1,6 +1,6 @@
 -- A function that divides two numbers
 DELIMITER $$
-CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS DECIMAL
+CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS FLOAT
 BEGIN
   IF b = 0 THEN
     RETURN 0;
